Using unsigned ints where appropriate
Unsigned Integers offer some extra safety because we often do want
negative values. cxxopts even does most of the checking for us now. That
is pretty neat.

configure_command_line_parser moved from utility namespace in global.h
to main_helper.h

Added plausibility checks for Complex Plane parameters

+++ src/mandelparams.h
@@ -22,27 +22,29 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include <string>
 
 struct MandelParameters {
-    int xrange;
+    unsigned int xrange;
     double xdelta;
     double x;
     double xl;
     double xh;
 
-    int yrange;
+    unsigned int yrange;
     double ydelta;
     double y;
     double yl;
     double yh;
 
-    int bailout;
+    unsigned int bailout;
 
-    int zoom;
+    unsigned int zoom;
 
     std::string image_base;
 
     MandelParameters() {}
-    MandelParameters(int xrange, double xl, double xh, int yrange, double yl,
-                     double yh, double bailout, int zoom, std::string image_base)
+    MandelParameters(unsigned int xrange, double xl, double xh,
+                     unsigned int yrange, double yl, double yh,
+                     unsigned int bailout, unsigned int zoom,
+                     std::string image_base)
         : xrange(xrange),
           xl(xl),
           xh(xh),
